+++ ts/features/payments/home/screens/__tests__/PaymentsHomeScreen.test.tsx
@@ -179,7 +179,7 @@ const renderComponent = ({
         home: {
           shouldShowAddMethodsBanner
         },
-        bizEventsTransaction: {
+        receipt: {
           latestTransactions: transactions
         }
       }