+++ ts/features/pushNotifications/store/reducers/environment.ts
@@ -1,6 +1,7 @@
 import { getType } from "typesafe-actions";
 import { Action } from "../../../../store/actions/types";
 import {
+  setEngagementScreenShown,
   setPushPermissionsRequestDuration,
   updateSystemNotificationsEnabled
 } from "../actions/environment";
@@ -13,13 +14,15 @@ export type EnvironmentState = {
   onboardingInstructionsShown: boolean;
   systemNotificationsEnabled: boolean;
   pushNotificationPermissionsRequestDuration?: number;
+  engagementScreenShownThisSession: boolean;
 };
 
 export const INITIAL_STATE = {
   applicationInitialized: false,
   onboardingInstructionsShown: false,
   systemNotificationsEnabled: false,
-  pushNotificationPermissionsRequestDuration: undefined
+  pushNotificationPermissionsRequestDuration: undefined,
+  engagementScreenShownThisSession: false
 };
 
 export const environmentReducer = (
@@ -37,6 +40,11 @@ export const environmentReducer = (
       return {
         ...state,
         pushNotificationPermissionsRequestDuration: action.payload
+      };
+    case getType(setEngagementScreenShown):
+      return {
+        ...state,
+        engagementScreenShownThisSession: true
       };
   }
   return state;